10411 Customer specific macro replacements for reporting
Within the customer properties there is now the option to define individual
replacement macros to use in reports. It is possible to set up an arbitrary
number of macros per customer.

Per default, two macros are initiated $CUSTOMER_ID$ and $CUSTOMER_NAME$,
corresponding to the customer properties ID and Name.

CMK-2472

Change-Id: I5f97c008e745d8426956dbbefc658c42f4886145