10475 FIX Azure Checkplugins: Go to stale state if info is missing
Previously a service resulting from one of the Azure checkplugins went to
UNKOWN if the corresponding data was not found in the agent output, as is
common for many checkmk services.
Unfortunately, this happens too often to be usefull, as the Azure API appears
to not allways report the full data. Now the affected services go to a stale
state instead.

Change-Id: I880dc9087e0c6380edc08c898aa9b78a0dea0c17