10413 FIX Sites can not be assigned to global scope
Users can be assigned to configured customers and the provider or to a
global scope to have access to all sites. The same option generators is
used for passwords and sites.

In the case of users and passwords, this objects are available for a
customer or globally. For site the global scope does not make sense as the
site is not available globally, but assigned to a global customer which
doesn't exist. Thus this option is removed to avoid misconfiguration.

Customer valuespec cannot be repeated so easily

Change-Id: Ibfbb85e10ce6644ada17acf84264d17f5be65107