Cleanup special skel form of install targets
The skel targets were made to install files to OMD_ROOT/skel.  There was
not much difference between the real install targets and the skel
targets and there were always confusions about when to use which of
them.

We now simplify this and merge the skel targets into the install
targets. The [pkg]/skel directories are still handled by the global OMD
packaging mechanism.

Change-Id: Ifaca9c8e9e5b4c446eb299d530f47753941f57dd