Simplify cleanup of OMD packages
Since the build stuff is now centralized in dedicated directories, the
cleanup mechanic is now a simple removal of some base directory and
there is no need to have per package cleanup targets with individual
logic anymore.

For convenience we could add package specific cleanup targets in a
central place that cleanup only the package related stuff in a
standardized way.

Change-Id: I74c5288cefd8510720c2e0305d7c30f3ece75cec