10463 FIX cisco_power: discover unique items in one edge case
For some sensor descriptions the cisco_power check discovered items
that were not unique. One description was e.g. "Sw1, PSA Normal"
which was translated into the item "Sw1". Another sensor that used
the description "Sw1, PSB Normal" was then aliased to "Sw1" as well.
Such items can now be differentiated. The example sensors would now
be mapped to the items "Sw1 PSA {ID}" and "Sw1 PSB {ID}" where {ID}
is the ID of the sensor.

The affected services have to be rediscovered.

SUP-1192

Change-Id: I3895b355595ae69171daa9eae003f98e5c0b783c