Skip parsing second "Personalities" to make md check cluster aware
When using the checkmk cluster feature, the output of two <<<md>>> sections is
merged. The second line can falsely be parsed as output of the last raid block.
Hence, when looking for working disks, we need to skip lines starting with:
Personalities
Otherwise, the check would return CRIT for the last md block of the first
cluster member because it parses e.g. "raid1" as a working disk, which then
doesn't list the right number of working disks in line 192 where active disks
are compared to the number of "U" in working_disks