10558 FIX Extension packages: Fixed performance issue
The performance of the calculation of the packaged files has been
improved to make the extension package management faster. Especially
when managing a larger number of files, this change can improve the user
experience of the package management.

In addition to the performance fix made in 1.6, the overall performance
of the WATO Extension packages module has been improved by changing some
things in the underlying library layer.

Change-Id: I38a646fe10b00a3fb96d8613b9d2a0cbcb9b9fb1