mk_oracle: removed ts_quotas from ASYNC_SECTION
The SQL for ts_quotas has beend added and enabled a long time ago but checkcode
is still missing. The SQL need a redesign before a check code be added. Due to
the problem with the SQL and missing checkcode the Section is removed from
ASYNC_SECTIONS.
It will be fixed and reenabled at a later time