10627 esx_vsphere_counters.cpu: Add new CPU usage check for ESX host systems
You can now monitor the CPU ready and Co-Stop values for ESX host systems. The values are the percentage of time in a check interval that vCPU(s) are ready but are no physical CPU is free and assigned to them.
VMWare calls this check interval 'realtime' which by default is equal to 20 seconds. The recommended VMWare levels are 5% for warning and 10% for critical.

Change-Id: I8e68382c1a0e1989dca1da7f8208683b75ecc83e