10643 redis_info: New check to monitor Redis instances
With this check you can monitor Redis instances. The check gets input from the
redis-cli command "info" and the resulting "Server" section. It outputs the
mode the server is running ("standalone", "sentinel" or "cluster") and the
uptime. You can set levels for both values. Furthermore, it outputs the server
version, the GCC compiler version, the PID, IP and port the server is listening
on.

Needs the agent plugin "mk_redis" to be installed.

Change-Id: If4d3b191b8fb06282992f9af2108e14a79b2376c