10644 graylog_sources: New check to monitor sources of a Graylog instance
This check monitors the count of messages for each source of a Graylog
instance. It outputs the total number of messages and the average rate of
messages in a given timeframe. You can choose if the service is shown on the
queried Graylog host or the sidecar host.

Needs the datasource "Check state of Graylog" to be configured.

Change-Id: Ia650e2fc19122148d3ad389cb97834e3811f53a3