Add dedicated mypy target to check Python 3 files
The test-mypy target is currently only testing in Python 2
and misses a lot of issues. We should maintain two independent
jobs, just like we do with Pylint.

One job checks Python 2 files in Python 2 mode and one job the
files that are relevant for Python 3.

The Python 3 mode uses the "test-py3/find-python-files" script that is
meant centralize the file discovery for all test jobs.

Change-Id: Ice2b2b4c1539d6f0e25ec699bf6904fed7dff6c0