Revert to previous site.version API in case a version can not be determined
The SiteContext.version property needs to deal with the case that a version of a site
can not be determined, e.g. because the version link is broken. Previous to the recent
refactorings this function returned a Optional[str] which has been changed to raise
an exception in the moment the version information is missing. This was not really
cleaning up the situation and has been reverted now.

Change-Id: I2b19dadb5be0e21c15b376f6d2e3704aee103891