10514 FIX mk_oracle.ps1: Fixed version check for instance services
The plugin tried to use the extended output for instance services even if
the version was not capable to produce the output. This happend, because
the version check is on Windows more specific than in Linux and the version
check was too loose. The plugin needs to be replaced and then the correct
output will be generated.

Change-Id: Iec29d77c275dcadf936fab1dcb934ffe8881720d