Get rid of pointer arithmetic (1)
This patch introduces a new Column type: `IntLambdaColumn` that takes a
functor taking no arguments and returning an int.  The class does not
take Column::Offsets.

This new class could, in principle, replace the children of `IntColumn`
For now, it lets us remove the most trivial ones:

 - FixedIntColumn
 - IntPointerColumn
 - AtomicInt32PointerColumn

This class does *not* access the values via pointer arithmetic
and is safer.

Note:

It cannot be used for columns that use pointer arithmetic as a
JOIN-ersatz.  The JOIN-ersatz is an entirely different story and shall
be dealt with separately.

Change-Id: I49656213a3db45dab7e29e1c3b77504e56a23d22