10513 FIX mk_oracle: removed ts_quotas from plugin output
The SQL for ts_quotas has beend added and enabled a long time ago but checkcode
is still missing. The SQL need a redesign before a check code be added. Due to
the problem with the SQL and missing checkcode the Section is removed from
ASYNC_SECTIONS and will not be executed anymore.
It will be fixed and reenabled at a later time.

Change-Id: I6b261d18c0cb33b38ae6f8c1dfaea2de27e9dd83