10123 FIX Fixed race condition when changing host check command.
When the kind of a host check command was changed, e.g. from an active check
to smart ping, there was a small time window for scheduling to go wrong.
This resulted in endless log lines of the form "invalid command line '@foo'
for bar", and the only way to resume normal operation was to restart the
Checkmk Micro Core, reloading was not enough.  This has been fixed.

Change-Id: I056880380bf61ff3e0ea0a1c5f84a8bca01843c4