Rewrite TableContacts without pointer arithmetic
Introduce `Table::IRow`: Row::rawData<>() returns the data with a
static_cast and static_cast only safely returns the type that was passed
to the void*.  This type is now `Table::IRow`.  Note that `Table::IRow`
is dumb and must be dynamically typed to something useful.

This patch adds a few more `*LambdaColumn`s that do not
use `Column::Offsets`.

The AttributeListLambdaColumn should probably be refactored into a more
generic dict column.  There already is a `dict` type for the columns
anyway.  However, this is not necessary now and may be done when the
tables do not rely on pointer arithmetic anymore.

CMK-3483

Change-Id: Iace5121d9e4030b1b97daae4d34149fccb9ea954