Use self.role_ids instead of user.role_ids in LoggedInUser
This way we're sure to always use the roles of the user object
itself. Since the cmk.gui.config.user object is usually set to
the current user this did not produce any bugs. We would only
end up with wrong permissions if the used LoggedInUser object
is different from the cmk.gui.config.user object.

CMK-3464

Change-Id: I7b72a6a1c2f31e02f5961387a70b5c515b2fe8e1