Addition to #10547: Improve existing view link dashlet creation
When adding a new dashlet, which should refer to an existing view,
a dedicated dialog for selecting the view is needed. This is necessary
because the "dashlet edit" dialog needs to know the datasource and
infos of a view before it can be used to edit a linked view dashlet

Now reusing the dialog which is used for cloning existing views as
dashlets.

Change-Id: Id27b74022a5552866a97e6d7f47bfa5250ceccac