Fix failing integration tests
The host and service checks for the integration test
test_simple_rbn_notification were disabled before the
switch of the core and were therefore not effective in
the first test for the Nagios core. This may lead to
failing integration tests if a host check is executed
with status UP while a host check with status DOWN is
expected by the test inside the function call
send_host_check_result.

Change-Id: Ifda9945fafb6af863f79538ba1727dffde860080