10630 graphs/perfometers: Improve graph colors to match the color themes
The color of graphs and perfometers are sometimes too bright or too dark compared to the rest of the interface, depending on the user interface theme that was used (classic, modern, dark).
The opacity can now be defined. Added support of the 'rgba' notation for the color of graphs. (rgba: '#aabbccff')

Change-Id: If8322c5016d1b732094f0cc1af64222811a5c35d