10766 FIX Fix custom graphs designer fraction
Custom graphs allows to compose metrics through an operation. Fraction,
that is the ratio between 2 metrics, was not recognized by the GUI
editor. This werk fixes the issue.

Because we use the application of operations as submit button, then the
operator goes as a variable name and not as a value in the request. We name
division with its mathematical symbol "/", and thus slashes need to be
accepted in variable names.

Change-Id: I3b195bbfe9c360ccd1cc9e3e71a3fe0810cc9d02