10618 New versioning of CMC configs
The config files to CMC are now versioned.

When the core detects an incompatible version, it
provides instructions to the user to generate a
new configuration instead of crashing with a
segmentation fault.

The new versioning scheme is consistent with the
versioning of the state files.

<b>Note:</b>

This patch <b>requires</b> that the configuration be
generated once more with <tt>cmk -U</tt>.

CMK-366

Change-Id: Ic74aa652c246fc443db36798add88a322259f6fe