10758 redis_info.persistence: New check to monitor Redis persistence
With this check you can monitor Redis instances. The check gets input from
the redis-cli command "info" and the resulting "Persistence" section. It
outputs the state of the last RDB (Redis Database Backup) save and AOF (Append
Only File) rewrite operation. Furthermore the time of the last successful RDB
save and the number of changes since the last dump.

Needs the agent plugin "mk_redis" to be installed.

Change-Id: I06acee6903a5fd8a182962c58bab9d14905f5bf8