10432 SEC cmk-update-agent.exe: Fix security issue on Windows
Recently, a vulnerability of PyInstaller, that we use to compile
the cmk-update-agent.exe executable on windows to one file, has been
discovered, see
<a href=https://github.com/pyinstaller/pyinstaller/security/advisories/GHSA-7fcj-pq9j-wh2r>here</a>

Only the windows version of the cmk-update-agent binary is affected. Unix versions and the python script
are not affected.

We fix this issue by updating to PyIntaller 3.6.

Jira: CMK-3565

Change-Id: Ia3774a8b29057e8d56df7678c3f9afeb39a751f4