10560 Extension packages: Add informational "Valid until version" field
Checkmk extension packages (MKPs) can now have a new field, which is named
"Valid until Checkmk version". A package provider can now declare that a
package is useful until a specific Checkmmk version. This may be useful when
backporting fixes or features to older versions where it is already known that
the official Checkmk package ships the same things as the package.

Please note that this field is only for informational purposes at the moment.
There is no automatic disabling of extension packages based on this field.

Change-Id: I90dea4e5d1e595c2c488623c9a94f1f33b65469c