10588 FIX ps: Handling of non-ASCII characters during process discovery
This fixes a rare encoding related exception during process discovery.

If one of the matched groups contained a non-ASCII character and that
matched group was supposed to be formatted into the service description,
the discovery crashed with an UnicodeEncodeError.

Users affected by this issue can open the corresponding WATO rule
"Process Discovery" and re-save it, to solve the issue.

Change-Id: Ief36ffccbbbfec9e30cca8e0580d9ee8616bf80a