10739 FIX send notifications chronologically
Due to a bug in the notification spooler new spoolfiles were processed before
old spoolfiles. This may lead to suprising results. If e.g. a host goes DOWN
and UP in a short time interval the UP notification may arrive before the
DOWN notification.

Now the notification spooler processes the notifications in the order of
occurance, i.e. the DOWN notification will be sent before the UP notification.

SUP-1708

Change-Id: Iee0e0ccc5670eda9108cab9cd370b1558d16cd11