10768 Report front matter templates
Some reports would benefit from a front matter to introduce the
content. For this purpose a report layout can now be transformed into a
front matter layout. In this case the layout is exclusively dedicated to
describe the front matter of the report, like cover page, table of contents
and other static elements one wishes the report to include.

Reports wishing to use the front matter template need simply to select it
in the {{report properties}}->{{Use settings and page layout from}}. The
front matter always generates an even number of pages and is numbered with
roman numerals.

The Table of contents report element(werk #10409) is only active in the
front matter templates.

CMK-3404

Change-Id: I04e1a5e25d7e1336beba124f025f9761b97d9ca6