Some minor refactorings on the way to fix hanging pylint
The current state of affairs is that pylint hangs for ever when going to
check the notify.py. During the recent typing additions some "return None"
statements have been added to make mypy happy. This seems to be a problem
for pylint.

The problem is currently being narrowed down with printf-debugging in the
pylint checkers.

The changes in this commit did not fix the bug, but are still good
cleanups.

Change-Id: I2ee0e8d69fa0faca8b018f83a948b40e8a168bcf