Setting CSP header in expected format
Previous releases used the "," separator to concatenate the different
parts of the content security policy header. This was done using the
apache mod_headers "append" header mechanism. The main reason to do so
was that this is more readable in the apache configuration than setting
a header with a single long value.

Sadly this "," separator is not used as separator of the CSP header
which makes browsers misunderstand the configuration.

We now set a single header with ";" as separator which makes the
browsers understand the policy as intended.

Change-Id: Ide0207908c4f84d8decf0bc4e0a91a030d80b3bf