Partially revert the cStringIO changes, trying to unbreak our integration tests.
io.StringIO is *not* a drop-in replacement for cStringIO.StringIO, the
latter does an implicit conversion from unicode to str.  Furthermore,
sys.stdout and friends have type IO[str] in both Python 2 and 3, so neither
io.BytesIO nor io.StringIO is a portable solution here.  We'll have to
figure out how to fix things here later.

Change-Id: I82f4e33511cf2622dd963018a6a5e9397349e9ae