10741 FIX lparstat_aix.cpu_util: display the correct physical CPU consumption
The physical CPU consumption for uncapped partitions is now taken directly
from the output of the lparstat command. Previously it was calculated based
on the CPU utilization which yields wrong values for uncapped partitions
if the physical CPU consumption is above the entitlement. Since the CPU
utilization is calculated with respect to the available CPUs the displayed
value never exceeded the entitlement.

Change-Id: If5e56c8b07c85735d42c000357638acf24730f5c