Render the sidebar in the main page and use an iframe for the content
If the sidebar is enabled it is now rendered directly in the HTML
page. The deprecated frameset that rendered the sidebar and the
site content in different frames is removed. The site content is
now rendered in a separate iframe. This is necessary since other
sites like e.g. the DokuWiki are displayed inside this iframe as
well.

To make this transition work different places had to be adapted:

* The sidebar is now rendered in cmk/gui/main.py.
* The method show of SidebarRenderer now receives the to optional
  arguments title and content.
* The function reset_sidebar_scheduler is introduced in sidebar.js
  to schedule an immediate sidebar reload.
* The new AJAX endpoint sidebar_ajax_add_snapin is extracted from
  page_add_snapin. The JS counterpart is the function add_snapin
  in sidebar.js.
* The sidebar is now minimized via the property visibility.
* References to frames[1] of a window were changed to frames[0]
  in various places.
* The CSS styling was adapted in various places.

Change-Id: I3c578ff484f697495cb6f23cc680fb73164b5728