10599 FIX Network interfaces: Handle illegal description/alias characters differently
The removal of illegal characters from interface descriptions and aliases is now handled consistently.

Some services may go to UNKNOWN for users who have configured the use of descriptions or aliases in
the service description.
This will only happen if the interface desription or alias did indeed contain one of the
characters <pre>`;~!$%^&*|\'"<>?,()=</pre>.

If this is the case, please perform a rediscovery. For every affected interface you will be informed
of one vanished an one new service by the same name. Performance data will not be lost.

Potentially affected plugins are

<ul>
<li>if_brocade</li>
<li>statgrab_net</li>
<li>if64adm</li>
<li>fritz</li>
<li>hp_msa_if</li>
<li>hitachi_hnas_fc_if</li>
<li>hpux_if</li>
<li>if64_tplink</li>
<li>if_lancom</li>
<li>esx_vsphere_counters</li>
<li>emc_vplex_if</li>
<li>mcdata_fcport</li>
<li>if64</li>
<li>ucs_bladecenter_if</li>
<li>lnx_if</li>
<li>if_fortigate</li>
<li>if</li>
<li>vms_if</li>
<li>aix_if</li>
<li>netapp_api_if</li>
<li>winperf_if</li>
</ul>

Change-Id: I6d2559fad091174ab31c2a9c1b4f178b4b143385