API: extract agent section
We use the aggregated information from check_info, snmp_scan_functions
and snmp_info in cmk.base.config to create section objects compliant
to the new API requirements of cmk.base.api.agent_based.

The resulting values from the 'registered_{snmp,agent}_sections'
dictionaries are then used when
 * scanning for SNMP devices
 * fetching the SNMP data
 * parsing the raw sections
 * discovering HostLabels

A lot of cleanup remains to be done once we have a similar conversion
for check- and inventory plugins in place.

CMK-4005

Change-Id: I7fd6df3292f9f47508902bb9f448215852298940