API: extract agent section
We use the aggregated information from check_info, snmp_scan_functions
and snmp_info in cmk.base.config to create section objects compiant
to the new API requirements of cmk.base.api.agent_based.

The resulting values from the 'registered_{snmp,agent}_sections'
dictionaries are then used when
 * scanning for snmp devices
 * fetching the snmp data
 * parsing the raw sections
 * discovering HostLabels

A lot of cleanup remains to be done once we have a similar conversion
for check- and inventory plugins in place.

CMK-4005

Change-Id: I47386bc74235f7cb09e0f624d3738570288c5a48