10859 agent_prometheus: addition of kube-state-metrics compatibility
This change introduces the kube-state-metrics package into the Prometheus
agent. It allows to monitor the connected Kubernetes cluster and associating
entities through the Prometheus connection. This monitoring option along with
selection of the entity levels can be configured in the Prometheus Datasource
Programs ruleset. The queried information uses the same k8 check-plugins as
the native Kubernetes agent. Information about the cluster, nodes, pods, daemon
sets and services can be retrieved through this option.

Change-Id: I6e4c0f218f9c39d92a8280bed2bf6433d466f5f8