10863 FIX Fix different issues with hosts created via Web API using JSON
When hosts are created using the Web API with data structures in JSON format,
configuring these hosts could lead to problems in different situations.

One example is that configuration rules that are created explicitly for such
a host can not be loaded for editing. One often used feature is creating
"ad hoc" disabled services rules on the service discovery page. These rules
lead to an error message "Unable to read current conditions of this rule. Falling back to default values."
when trying to edit the created rule.

This issue has been fixed. Starting with the 1.6.0p10 the hosts are initially
correctly saved. Existing host and rule configurations will be fixed during
the update to 1.6.0p10.

Change-Id: Ic71a849e6f7c23938a46beda1668baede45a8542