10865 FIX Fix additional rule analyzation issue on service object parameter page
Werk #10696 and #10865, which were shipped with 1.6.0p8 and 1.6.0p9, already
fixed issues on the "service object parameter page" that lead to wrong
effective settings being displayed on the page.

The changes were not enough to fix all problems. There was an additional issue
found and fixed that was affecting service related rulesets only.

Change-Id: I2c16cb739fb27b181e55ab9f0aafe29bdf94b52f