Use Python 3 YAPF for all our Python code
The Python 2 YAPF command is not able to deal with Python 3 style type
hints. Since we want to switch from comment based type hints to native
Python 3 type hint syntax one day this seems to be a good idea.

The only difference for Python 2 code is that builtins like "exec" are
formatted differently with by the Python 3 YAPF.

Change-Id: I01704523614d52bf0f2e0080e7f52a8a0f4bd7cf