10435 Agent Deployment: Log info about calling host on error
When calling the agent bakery for the purpose of agent deployment
(normally done automatically via agent updater), occuring errors will be logged
to ~/var/log/web.log. This error log entry is now augmented by the calling
hostname (if it is provided by the calling client). With this information, it
becomes easier to debug host-related errors.

CMK-3801

Change-Id: I480b9e445c674a34c1596cf28b7ad20c0e325ea1