10867 FIX Tags: Fix cleanup actions after tag choice removal / editing
When removing a tag choice or changing the ID of a tag choice, this leads
to a warning message when e.g. rulesets refer to the modified tag choice.
After choosing one of these actions and confirming the dialog the action
was not performed as expected. This is now possible again.

Change-Id: I75d9f0ef23ce4d7d596a3ab921bb89ac8c083e51