10868 FIX Tags: Custom "Agent type" tag group cleanup is now easier
Werk #5535 introduced with Checkmk moved the "Agent type" tag group from
a tag group that could be customized to a builtin tag group.

In case you have modified it in any way, the tag group could not be migrated
manually. In this situation the new builtin tag choices have been added to
the tag group and the titles of the previously existing tag choices have been
prefixed with "Legacy: ".

To perform the cleanup now you have to perform the following steps:

<ul>
<li>Remove all tag choices except <tt>cmk-agent</tt>, <tt>all-agents</tt>,
    <tt>special-agents</tt> and <tt>no-agent</tt>.</li>
<li>Then you can try to delete the tag group.</li>
</ul>

If you have prepared the tag group as mentioned in the first step, the second
step should result in a message "Cleaned up user tag group to builtin". The
tag group is visible as "(builtin)" tag group.

Change-Id: I8de54593f0e631fafa88c7bcb68f1871af280042