10908 FIX snmp_uptime: Fix crash for some Fortigate devices
Users experienced crashes of uptime checks with some Fortigate devices. This is because
the data delivered by these devices does not contain any timestamps. As a solution, the
uptime check will not be discovered any more if the delivered data cannot be interpreted
as a timestamp. To apply this fix, users have to re-discover the services of the
corresponding hosts.

Change-Id: Iba3e52df1335c1cdb2fd724e7a58c56d3ac8364f