10872 FIX agent_vsphere: Removed ESX 4.1 compatibility mode
The ESX 4.1 compatibility mode of the ESX vSphere special agent has been
removed.  This separate implementation, which was based on the pysphere python
API, was outdated, much slower and missed several features of the current ESX
vSphere special agent.

In case you still need this, you may have to setup a separate site which uses
the old special agent to collect the needed information.

Change-Id: I391a5ee527100bfd930d29ba4569de5aa5910e0d