Unbreak alert/notification scripts.
They were damaged by the recent changes to the license headers.

The leading comments in these scripts have an actual meaning, see the
cmk.gui.watolib._load_user_scripts_from() function.

If we want to have the new license headers there, we have to either a) make
that function smarter or b) put the license header after the magic comments.

Change-Id: Ida49a3f9bc9c8bb9d3a1b86a18733fd73669ebd1