Compat pretty printer: Format unicode strings with hex escapes
When Python 3 creates a repr which is interpreted by Python 2, we need
to produce a repr-string where non ASCII characters are hex escaped as
Python 2 usually does.

This is needed to make e.g. integration tests work that send Python
requests to the Python 2 WebAPI (like the integration test
integration/core/test_active_checks.py does). If we don't do this, the
resulting string on the Python 2 side is not correctly encoded.

Change-Id: Ic400d07b08829a8d386b525a04ae6f1511e5a75c