Fix composition tests for the moment
They still need to run in Python 2, because they are loading the agent
updater as Python module, which is not Python 3 compatible yet.

This commit fixes some compatibility issue in the testlib that is used
to prepare sites for running the composition tests.

To be able to migrate the composition tests over to Python 3, we need to
port the agent updater to Python 3 first.

Some minor flake8 cleanups on the way.

Change-Id: I6930052eee04ec1c155659b3b2f8748a9402dd36