Simplify installation of Python 3 cmk module
Now that a lot of stuff has moved to Python 3 it is possible to
install the whole cmk module in a single step. Several files were
installed twice btw.

This additionally removes cmk/__init__.py and cmk/base/__init__.py
to make them namespace packages. This is needed to make the local
hierarchy shadowing work in the near future.

Some more cleanups:

* Exclude precompiled .pyc and __pycache__ files from check_mk
  "upstream" tar archive
* Add missing mkdir calls to make single check_mk package make target call work

Change-Id: I3e45c3c8b9dd195a8ec9fb35c6dcfceb2fadb7cf