10869 FIX Fix potential "Request-URI Too Long" error when editing huge tag groups
When editing a tag group with a larger number of tag choices an error "Request-URI Too Long"
could appear.

When editing a tag choice that is configured for some rules or hosts, a confirm dialog with
the tile "Your modifications affect some objects" is shown to the user that asks which action
should be taken on the related object. When choosing an action and submitting the form, the
error mentioned above could occur.

Change-Id: I14dd8589a0855b55ad4476905f404d73fea38bfa