10974 SEC Agent Bakery: require minimum length of 8 characters for agent signing key
A minimum length of 8 characters is now required for the password for newly generated agent signing
keys at the agent bakery.

Already existing keys or keys that get imported are not affected.

However, we would always suggest to choose a strong password, as the private
key of the agent signing certificate is encrypted using this very password, and a weak/short
combination could introduce a vulnerability.

CMK-3782

Change-Id: If88f16489d17a6488effa0bc7d58508b2cefac93