10948 FIX systemd_units: WATO rules modification to allow compability
In the previous version, the splitting of WATO rules for the systemd
individual services and the systemd summary service was performed
without compability for previously configured systemd rules.
This incompability led to a warning display within the WATO rule page.
This change resolves the warning display by ignoring the old, incompatible keys.

Change-Id: Ia85d4d8b2b2bfea397e2f83894ac6a82e12c94a3