10954 ucd_mem: count buffered and cached memory as free
For memory checks based on the output of a UCD SNMP deamon (check ucd_mem),
cached and buffered memory are now counted as free memory. This is consistent
with the behaviour of the agent for Linux systems. As a result, the memory
usage displayed by checkmk for hosts with this service might decrease if
buffered and cashed memory are part of the SNMP information given to checkmk.

Change-Id: I52e07d6aa930a8bf663b868326c05c8daecae313