Create directory structure for new plugin hierarchy
The concept here is (from a users point of view):

* The builtin files are located at "lib/check_mk/base/plugins/agent_based"
* Custom files can be placed at "local/lib/check_mk/base/plugins/agent_based"

To have a consistent experience it is important to have equal paths in
both domains, the builtin path and the local hierarchy.

The directory structure has been choosen to reflect the Checkmk internal
component separation. A path of "lib/check_mk/base/plugins/agent_based"
is built of these parts: "lib/[software]/[component]/plugins/[plugin-domain]".

We don't want to expose the "lib/python3" base directory to the user.
This is an internal path. For this reason we use the "lib/check_mk"
symlink to guide the user to the correct location.

In the long term we'll move all our plugin domains to this path structure.

Change-Id: Ifd075e0dd3f70d127dcc82228a0aa606c69a93d6